refactor(ext/node): consolidate node:fs modules (part 4)
Consolidate `_fs_rename.ts`, `_fs_rm.ts`, and `_fs_rmdir.ts` into
`fs.ts` and `internal/fs/promises.ts`, continuing the pattern from
parts 1-3.
