+++ ext/node/lib.rs
@@ -397,9 +397,6 @@ deno_core::extension!(deno_node,
     "_fs/_fs_readlink.ts",
     "_fs/_fs_readv.ts",
     "_fs/_fs_realpath.ts",
-    "_fs/_fs_rename.ts",
-    "_fs/_fs_rm.ts",
-    "_fs/_fs_rmdir.ts",
     "_fs/_fs_stat.ts",
     "_fs/_fs_statfs.ts",
     "_fs/_fs_symlink.ts",
